place Media to a separate file
