+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/livia/Desktop/magic-composition/src/Components/utils.ts":"1","/Users/livia/Desktop/magic-composition/src/Components/ImportModal/ImportModal.tsx":"2","/Users/livia/Desktop/magic-composition/src/Components/index.tsx":"3"},{"size":707,"mtime":1610598206405,"results":"4","hashOfConfig":"5"},{"size":1175,"mtime":1610598237892,"results":"6","hashOfConfig":"5"},{"size":421,"mtime":1610598255364,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"fzgybw",{"filePath":"10","messages":"11","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"12"},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/livia/Desktop/magic-composition/src/Components/utils.ts",[],"/Users/livia/Desktop/magic-composition/src/Components/ImportModal/ImportModal.tsx",["15"],"import React, { useState } from \"react\";\nimport { ViewMedia, DragModal } from \"./ViewMedia\";\nimport styles from \"./ImportModal.module.css\";\nimport { Button } from \"react-bootstrap\";\n\ntype Props = {\n  setShow: React.Dispatch<React.SetStateAction<boolean>>;\n  removeFile: (index: number) => void;\n  addFile: (newMedia: File[]) => void;\n};\n\nexport const ImportModal: React.FC<Props> = ({\n  setShow,\n  removeFile,\n  addFile,\n}) => {\n  const createMediaElement = (files: File[]) => {\n\n  }\n  return (\n    <div className={styles.canvaHomePage}>\n    <div className={styles.dropModal}>\n      <div\n        className={styles.dropModal}\n        onDragLeave={(e) => {\n          e.preventDefault();\n        }}\n      >\n        <DragModal />\n      </div>\n        <ViewMedia medias={[]} removeMedia={()=>{}} createMediaElement={createMediaElement}/>\n        <Button\n          className={styles.createVideoButton}\n          onClick={() => setShow(true)}\n          variant=\"secondary\"\n        >\n          Create Video 🎬\n        </Button>\n      </div>\n    <span> Or </span>\n    <span className={styles.desktopOnly}>\n      Drag &amp; Drop your files here 📥\n    </span>\n    </div>\n  \n  )\n};\n","/Users/livia/Desktop/magic-composition/src/Components/index.tsx",[],{"ruleId":"16","severity":1,"message":"17","line":1,"column":17,"nodeType":"18","messageId":"19","endLine":1,"endColumn":25},"@typescript-eslint/no-unused-vars","'useState' is defined but never used.","Identifier","unusedVar"]
+[{"/Users/livia/Desktop/magic-composition/src/Components/utils.ts":"1","/Users/livia/Desktop/magic-composition/src/Components/ImportModal/ImportModal.tsx":"2","/Users/livia/Desktop/magic-composition/src/Components/ImportModal/ViewMedia.tsx":"3","/Users/livia/Desktop/magic-composition/src/App.tsx":"4"},{"size":707,"mtime":1610671009901,"results":"5","hashOfConfig":"6"},{"size":1114,"mtime":1610671736611,"results":"7","hashOfConfig":"6"},{"size":1610,"mtime":1610671397589,"results":"8","hashOfConfig":"6"},{"size":807,"mtime":1610671533282,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"fzgybw",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/livia/Desktop/magic-composition/src/Components/utils.ts",[],"/Users/livia/Desktop/magic-composition/src/Components/ImportModal/ImportModal.tsx",[],"/Users/livia/Desktop/magic-composition/src/Components/ImportModal/ViewMedia.tsx",[],"/Users/livia/Desktop/magic-composition/src/App.tsx",["18"],{"ruleId":"19","severity":1,"message":"20","line":7,"column":10,"nodeType":"21","messageId":"22","endLine":7,"endColumn":14},"@typescript-eslint/no-unused-vars","'show' is assigned a value but never used.","Identifier","unusedVar"]
