+++ src/Components/ImportModal/ImportModal.tsx
@@ -1,10 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import { ViewMedia, DragModal } from "./ViewMedia";
 import styles from "./ImportModal.module.css";
 import { Button } from "react-bootstrap";
 
+
 type Props = {
-  setShow: React.Dispatch<React.SetStateAction<boolean>>;
+  setShow: (show: boolean) => void;
   removeFile: (index: number) => void;
   addFile: (newMedia: File[]) => void;
 };
@@ -14,21 +15,21 @@ export const ImportModal: React.FC<Props> = ({
   removeFile,
   addFile,
 }) => {
-  const createMediaElement = (files: File[]) => {
+  //TODO
+  const createMediaElement = () => {
 
   }
+
   return (
     <div className={styles.canvaHomePage}>
     <div className={styles.dropModal}>
       <div
         className={styles.dropModal}
-        onDragLeave={(e) => {
-          e.preventDefault();
-        }}
+        onDragLeave={(e) => {e.preventDefault()}}
       >
         <DragModal />
       </div>
-        <ViewMedia medias={[]} removeMedia={()=>{}} createMediaElement={createMediaElement}/>
+        <ViewMedia medias={[]} removeMedia={removeFile} createMediaElement={createMediaElement}/>
         <Button
           className={styles.createVideoButton}
           onClick={() => setShow(true)}
@@ -37,9 +38,9 @@ export const ImportModal: React.FC<Props> = ({
           Create Video 🎬
         </Button>
       </div>
-    <span> Or </span>
+    <span>Or</span>
     <span className={styles.desktopOnly}>
-      Drag &amp; Drop your files here 📥
+      Drag & Drop your files here 📥
     </span>
     </div>
