--- src/Components/index.tsx
@@ -1,14 +1,0 @@
-export { ImportModal } from "./ImportModal/ImportModal";
-export { Header } from "./Shared/Header";
-export { isImage, isVideo, trimmedName, imageDuration, audioSound } from "./utils";
-
-export class Media {
-  filename: string;
-  type: string;
-  element: JSX.Element;
-  constructor(filename: string, type: string, element: JSX.Element) {
-    this.filename = filename;
-    this.type = type;
-    this.element = element;
-  }
-}
